AAU-BSc-Software/Project-P6/backend/requirements.txt
aenum==3.1.16
asn1tools==0.167.0
asttokens==3.0.0
bidict==0.23.1
bitstruct==8.20.0
black==25.1.0
boltons==25.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorlog==6.9.0
contourpy==1.3.2
coverage==7.8.0
cycler==0.12.1
dacite==1.9.2
decorator==5.2.1
dnspython==2.7.0
eventlet==0.39.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
future==1.0.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.2.1
h11==0.16.0
idna==3.10
iniconfig==2.1.0
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
parrot-olympe==7.7.5
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
protobuf==3.19.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycryptodomex==3.22.0
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
shapely==2.1.0
simple-websocket==1.1.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tomli==2.2.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==3.0
ultralytics==8.3.127
ultralytics-thop==2.0.14
urllib3==2.4.0
wcwidth==0.2.13
wsproto==1.2.0
